whatsapp chat analyzer app/requirements.txt
emoji
fpdf2
matplotlib
pandas
plotly
python-dateutil
seaborn
streamlit
textblob
vaderSentiment
wordcloud
